Began adding features to determine if in loop.
Decided the proper course of action was to "disassemble" to source notes
when first loaded.  I can't seem to find a way to jump directly to where I
need based entirely on the pc.

A possible future optimization to keep in mind would be to build the Notes
object on the fly during execution.  Then again, on the fly would probably
incur costs throught the entire execution even after full disassembly.
Upfront disassembly is just a one time cost.